01-ChatBot-with-memory/requirements.txt
# LLM
langchain
langchain-openai
langchain-anthropic

# MongoDB
langchain-mongodb
pymongo

# Callbacks & Observability
langchain-community
langsmith

# Environment
python-dotenv

# UI
streamlit

# Encryption

cryptography
